+++ internal/guest/linux/doc.go
@@ -1,0 +1,2 @@
+// Package linux contains definitions required for making a linux ioctl.
+package linux
