+++ internal/guest/storage/vmbus/doc.go
@@ -1,0 +1,1 @@
+package vmbus
